requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
langchain
openai==1.3.5
azure-storage-blob
azure_search_documents
PyPDF2==3.0.1